smoothixir v0.1.0 Smoothixir

Documentation for Smoothixir. Smoothixir is a scrapper module of the website allrecipes.com it aims at creating a list of smoothies with their name, ingredients and directions.

Functions

display_smoothies(arg)

Simple function that displays smoothies in a nice way in the console.

Examples

iex> Smoothixir.get_smoothies_recipe() |> Smoothixir.display_smoothies()
Triple Threat Fruit Smoothie
1 kiwi, sliced1 banana, peeled and chopped1/2 cup blueberries1 cup strawberries1 cup ice cubes1/2 cup orange juice1 (8 ounce) container peach yogurt
In a blender, blend the kiwi, banana, blueberries, strawberries, ice, orange juice, and yogurt until smooth.

+++++++++++++++++++++++++++++
Groovie Smoothie
2 small bananas, broken into chunks1 cup frozen unsweetened strawberries1 (8 ounce) container vanilla low-fat yogurt3/4 cup milk
In a blender, combine bananas, frozen strawberries, yogurt and milk. Blend until smooth. Pour into glasses and serve.

get_smoothie_directions(body)

Allow you to find the directions of a smoothie within a page’s html body. returns an Array of Strings.

Examples

iex> Smoothixir.get_smoothie_directions(body)
["In a blender, combine strawberries, orange juice, ice cubes and sugar. Blend until smooth. Pour into glasses and serve."]

get_smoothie_ingredients(body)

Allow you to find the ingredients of a smoothie within a page’s html body. returns a List of strings.

Examples

iex> Smoothixir.get_smoothie_ingredients(body)
["1 cup chopped fresh strawberries", "1 cup orange juice", "10 cubes ice", "1 tablespoon sugar"]

get_smoothie_name(body)

Allow you to find the name of a smoothie within a page’s html body. returns a String.

Examples

iex> Smoothixir.get_smoothie_name(body)
"Mongolian Strawberry-Orange Juice Smoothie"

get_smoothies_html_body(arg)

once you have an array of all the URLs of smoothies pages, this function will get the HTML body of each of the page.

Examples

iex> Smoothixir.get_smoothies_url() |> Smoothixir.get_smoothies_html_body
["<!DOCTYPE html><html lang="en-us"><head><title>Kiwi Strawberry Smoothie Recipe - Allrecipes.com</title></head><body></body></html>"] # simplified for readability

get_smoothies_recipe()

This is the main function of the module. It gathers all the previous functions and outputs a List of Maps of smoothies with name, ingredients, directions for each of them.

Examples

iex> Smoothixir.get_smoothies_recipe()
 {:ok,
   [%{
      directions: ["Cut banana into small pieces and place into the bowl of a blender.  Add the soy milk, yogurt, flax seed meal, and honey.  Blend on lowest speed until smooth, about 5 seconds.  Gradually add the blueberries while continuing to blend on low.  Once the blueberries have been incorporated, increase speed, and blend to desired consistency.                            "],
      ingredients: ["1 frozen banana, thawed for 10 to 15 minutes",
       "1/2 cup vanilla soy milk", "1 cup vanilla fat-free yogurt",
       "1 1/2 teaspoons flax seed meal", "1 1/2 teaspoons honey",
       "2/3 cup frozen blueberries"],
      name: "Heavenly Blueberry Smoothie"
    }
  ]}

get_smoothies_url()

allows you to fetch all url of each smoothie receipe.

Examples

iex> Smoothixir.get_smoothies_url()
{;ok,
["https://www.allrecipes.com/recipe/77035/kiwi-strawberry-smoothie/",
  "https://www.allrecipes.com/recipe/20816/delicious-healthy-strawberry-shake/",
  "https://www.allrecipes.com/recipe/23530/orange-smoothie/",
  "https://www.allrecipes.com/recipe/32425/mongolian-strawberry-orange-juice-smoothie/",
  "https://www.allrecipes.com/recipe/100822/heavenly-blueberry-smoothie/"]}